refactor(server): kb-app 四个上帝类按职责拆解
拆分边界的依据是依赖的独占性而非行数——某组协作者只被某一段代码读，
那一段就是一个独立职责。据此拆出 7 个协作者，四个应用服务的构造依赖
由 31/25/22/21 降至 20/22/18/10：

- 检索线：RetrievalNodeAssembler 接手"胜出片段怎么描述"；
  MultimodalRetrievalService 补齐一处既有不对称——图路早有独立的
  GraphRetrievalService，多模态执行逻辑却内联在管线里
- 索引线：ChunkSplitter（切分与落库）、IndexTaskLifecycle（任务状态机）、
  VersionActivationHandler（激活及其四项后续）
- 导入/评测线：ChatSessionImporter（会话写成文档版本）、
  EvalCaseRunner（跑单个用例，不碰 mapper 因而可安全并发）

顺带清理 judgeOneCase/judgeAll 中只用于向下透传、从未被读取的参数 k；
ReleaseGateService.runGate 提取 submitGateRuns（88 → 65 行）。

补 IndexPipelineServiceTest 10 例（此前无直接单测，引用它的三个测试类
都只把它当 mock），覆盖复用回退、待确认暂停、失败状态归类与页策略标志
传递；变异验证注入 5 处缺陷全部转红。

行为零变更：全部为 Extract Class/Method，未改任何逻辑分支，无 schema
迁移、无配置键、无 API 变更。测试 1340 → 1350 全绿，246 个 Spring bean
构造注入依赖图无环。ARCHITECTURE 升 v2.8，订正 §3.5 中因本次重构失准的
"page 策略由管线直接分流"一句。
